requirements.txt
# streamlit==1.29.0
# transformers==4.35.0
# torch>=2.2.0
# pandas==2.1.1
# numpy==1.26.0
# scikit-learn==1.3.1
# requests==2.31.0
# python-dotenv==1.0.0



# Core dependencies
streamlit>=1.24.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0
nltk>=3.8.1
matplotlib>=3.7.0

# For dataset handling
datasets>=2.13.0

# Optional dependencies for enhanced functionality
wordcloud>=1.8.0

# transformers
# scikit-learn
# torch
# datasets
# streamlit
# openai
# streamlit==1.22.0
# torch==2.0.0
# transformers==4.28.1
# pandas==1.5.3
# numpy==1.24.3
# scikit-learn==1.2.2
# openai==0.27.8
# streamlit==1.29.0
# transformers==4.35.0
# torch==2.1.0
# pandas==2.1.1
# numpy==1.26.0
# scikit-learn==1.3.1
# requests==2.31.0
# python-dotenv
